Split out the terminals into separate classes
Each class knows how to run the applescript needed to set a color or
font name.

Each class's applescript is also tested with the `RUN_APPLESCRIPT=1 be
rake` command, so that makes it easier to find out if something went
wrong.

Now all that's left to do is create one hydra class that takes the
applescript from each class and creates a wrapper above it that uses
Applescript to find out which class it should call.